+++ data_adapter/structure.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import csv
 import re
 from collections import defaultdict, namedtuple
 
@@ -17,21 +18,26 @@ class StructureError(Exception):
     """Raised if structure is corrupted"""
 
 
-AdditionalParameter = namedtuple("AdditionalParameter", ("subject", "isAbout"))
-
-ADDITONAL_PARAMETERS = {
-    "onshore wind farm": [
-        AdditionalParameter("net capacity factor", "onshore"),
-    ]
-}
+Link = namedtuple("Link", ("linked_process", "parameter"))
 
 
-def get_additional_parameters(process: str):
-    if process not in ADDITONAL_PARAMETERS:
+def get_links(links_name: str):
+    link_filename = settings.STRUCTURES_DIR / f"{links_name}.csv"
+    with open(link_filename, "r", encoding="utf-8") as link_file:
+        link_csv = csv.DictReader(link_file, delimiter=";")
+        links = defaultdict(list)
+        for line in link_csv:
+            links[line["process"]].append(Link(line["table_name"], line["column_name"]))
+    return links
+
+
+def get_links_for_process(process: str, links_name: str):
+    links = get_links(links_name)
+    if process not in links:
         return {}
     parameters = defaultdict(list)
-    for parameter in ADDITONAL_PARAMETERS[process]:
-        parameters[parameter.subject].append(parameter.isAbout)
+    for parameter in links[process]:
+        parameters[parameter.linked_process].append(parameter.parameter)
     return parameters