+++ tests/test_structure.py
@@ -1,7 +1,7 @@
 import json
 import pathlib
 
-from data_adapter.structure import get_energy_structure
+from data_adapter import structure
 
 
 def test_get_energy_structure():
@@ -12,4 +12,21 @@ def test_get_energy_structure():
     ) as hardcoded_json:
         dict_expected = json.load(hardcoded_json)
 
-    assert get_energy_structure(structure="modex_example") == dict_expected
+    assert structure.get_energy_structure(structure="modex_example") == dict_expected
+
+
+def test_get_links():
+    links = structure.get_links("hack-a-thon_links")
+    assert len(links) == 3
+    assert "modex_tech_generator_gas" in links
+    assert len(links["modex_tech_generator_gas"]) == 2
+    assert isinstance(links["modex_tech_generator_gas"][0], structure.Link)
+    assert links["modex_tech_generator_gas"][0].linked_process == "modex_constraint"
+    assert links["modex_tech_generator_gas"][0].parameter == "emission_costs"
+
+
+def test_get_links_for_process():
+    links = structure.get_links_for_process("modex_tech_wind_turbine_onshore", "hack-a-thon_links")
+    assert len(links) == 2
+    assert links["modex_capacity_factor"] == ["onshore"]
+    assert links["modex_constraint"] == ["wacc"]